Server compiles with gRPC now
