+++ src/server/Utils.hpp
@@ -1,0 +1,4 @@
+#pragma once
+namespace mik {
+void createLogger();
+}
