+++ test/CMakeLists.txt
@@ -1,1 +1,4 @@
-add_subdirectory(client)
+
+if (BUILD_CLIENT)
+    add_subdirectory(client)
+endif()
